fly-by fixes while looking at #11623
- unrelated: also make modern build work again. The less it digresses in the 'churn' period the better.
- deduplicate
